AvailableDate entitisi yapıldı. Doctor ile AvailableDate arasında Many To One ilişkisi kuruldu.

+++ src/main/java/net/atmacacode/veterinarian/dao/AvailableDateRepo.java
@@ -3,5 +3,5 @@ package net.atmacacode.veterinarian.dao;
 import net.atmacacode.veterinarian.entities.AvailableDate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvailableDateRepo extends JpaRepository<AvailableDate, Integer> {
+public interface AvailableDateRepo extends JpaRepository<AvailableDate, Long> {
 }
